Crate algebra_core

Re-exports

pub use self::bytes::*;

pub use self::serialize::*;

pub use self::fields::*;

pub use self::biginteger::*;

pub use self::curves::*;

pub use self::groups::*;

pub use self::msm::*;

Modules

biginteger

bytes

curves

fields

groups

io
no-std io replacement

msm

prelude

serialize

Macros

field_new

impl_additive_ops_from_ref

impl_multiplicative_ops_from_ref

push_to_vec

to_bytes
Takes as input a sequence of structs, and converts them to a series of bytes. All traits that implement Bytes can be automatically converted to bytes in this manner.

Traits

Error

One
Defines a multiplicative identity element for Self.

ToConstraintField
Types that can be converted to a vector of F elements. Useful for specifying how public inputs to a constraint system should be represented inside that constraint system.

UniformRand

Zero
Defines an additive identity element for Self.

Functions

log2
Returns log2

test_rng
Should be used only for tests, not for any real world usage.