Trait algebra_core::biginteger::BigInteger

pub trait BigInteger: ToBytes + FromBytes + CanonicalSerialize + ConstantSerializedSize + CanonicalDeserialize + Copy + Clone + Debug + Default + Display + Eq + Ord + Send + Sized + Sync + 'static + UniformRand + AsMut<[u64]> + AsRef<[u64]> + From<u64> {
    const NUM_LIMBS: usize;

    fn add_nocarry(&mut self, other: &Self) -> bool;

    fn sub_noborrow(&mut self, other: &Self) -> bool;

    fn mul2(&mut self);

    fn muln(&mut self, amt: u32);

    fn div2(&mut self);

    fn divn(&mut self, amt: u32);

    fn is_odd(&self) -> bool;

    fn is_even(&self) -> bool;

    fn is_zero(&self) -> bool;

    fn num_bits(&self) -> u32;

    fn get_bit(&self, i: usize) -> bool;

    fn from_bits(bits: &[bool]) -> Self;

    fn to_bits(&self) -> Vec<bool>;

    fn find_wnaf(&self) -> Vec<i64>;

    fn write_le<W: Write>(&self, writer: &mut W) -> IoResult<()> { ... }

    fn read_le<R: Read>(&mut self, reader: &mut R) -> IoResult<()> { ... }
}

This defines a BigInteger, a smart wrapper around a sequence of u64 limbs, least-significant limb first.

Associated Constants

const NUM_LIMBS: usize

Number of limbs.

Loading content...
Required methods

fn add_nocarry(&mut self, other: &Self) -> bool

Add another representation to this one, returning the carry bit.

fn sub_noborrow(&mut self, other: &Self) -> bool

Subtract another representation from this one, returning the borrow bit.

fn mul2(&mut self)

Performs a leftwise bitshift of this number, effectively multiplying it by 2. Overflow is ignored.

fn muln(&mut self, amt: u32)

Performs a leftwise bitshift of this number by some amount.

fn div2(&mut self)

Performs a rightwise bitshift of this number, effectively dividing it by 2.

fn divn(&mut self, amt: u32)

Performs a rightwise bitshift of this number by some amount.

fn is_odd(&self) -> bool

Returns true iff this number is odd.

fn is_even(&self) -> bool

Returns true iff this number is even.

fn is_zero(&self) -> bool

Returns true iff this number is zero.

fn num_bits(&self) -> u32

Compute the number of bits needed to encode this number. Always a multiple of 64.

fn get_bit(&self, i: usize) -> bool

Compute the i-th bit of self.

fn from_bits(bits: &[bool]) -> Self

Returns the big integer representation of a given big endian boolean array.

fn to_bits(&self) -> Vec<bool>

Returns the bit representation in a big endian boolean array, without leading zeros.

fn find_wnaf(&self) -> Vec<i64>

Returns a vector for wnaf.

Loading content...
Provided methods

fn write_le<W: Write>(&self, writer: &mut W) -> IoResult<()>

Writes this BigInteger as a big endian integer. Always writes (num_bits / 8) bytes.

fn read_le<R: Read>(&mut self, reader: &mut R) -> IoResult<()>

Reads a big endian integer occupying (num_bits / 8) bytes into this representation.

Loading content...