Module algebra_core::curves

Re-exports

pub use self::models::*;

Modules

models

Traits

AffineCurve
Affine representation of an elliptic curve point guaranteed to be in the correct prime order subgroup.

CycleEngine
A cycle of pairing-friendly elliptic curves.

PairingEngine

ProjectiveCurve
Projective representation of an elliptic curve point guaranteed to be in the correct prime order subgroup.

Functions

prepare_g1
Preprocess a G1 element for use in a pairing.

prepare_g2
Preprocess a G2 element for use in a pairing.