Trait algebra_core::fields::PrimeField

pub trait PrimeField: FftField<FftParams = Self::Params> + FromStr + From<Self::BigInt> + Into<Self::BigInt> {
    type Params: FpParameters<BigInt = Self::BigInt>;
    type BigInt: BigInteger;
    fn from_repr(repr: Self::BigInt) -> Option<Self>;

    fn into_repr(&self) -> Self::BigInt;

    fn qnr_to_t() -> Self { ... }

    fn size_in_bits() -> usize { ... }

    fn trace() -> Self::BigInt { ... }

    fn trace_minus_one_div_two() -> Self::BigInt { ... }

    fn modulus_minus_one_div_two() -> Self::BigInt { ... }
}

The interface for a prime field.

Associated Types

type Params: FpParameters<BigInt = Self::BigInt>

type BigInt: BigInteger

Loading content...
Required methods

fn from_repr(repr: Self::BigInt) -> Option<Self>

Returns a prime field element from its underlying representation.

fn into_repr(&self) -> Self::BigInt

Returns the underlying representation of the prime field element.

Loading content...
Provided methods

fn qnr_to_t() -> Self

Return the a QNR^T

fn size_in_bits() -> usize

Returns the field size in bits.

fn trace() -> Self::BigInt

Returns the trace.

fn trace_minus_one_div_two() -> Self::BigInt

Returns the trace minus one divided by two.

fn modulus_minus_one_div_two() -> Self::BigInt

Returns the modulus minus one divided by two.

Loading content...