Module algebra_core::serialize

Re-exports

pub use crate::bytes::FromBytes;

pub use crate::bytes::ToBytes;

pub use crate::io::Read;

pub use crate::io::Write;

Structs

EmptyFlags
Flags to be encoded into the serialization.

Enums

EdwardsFlags
Flags to be encoded into the serialization. The default flags (empty) should not change the binary representation.

SWFlags
Flags to be encoded into the serialization. The default flags (empty) should not change the binary representation.

SerializationError
This is an error that could occur during serialization

Traits

CanonicalDeserialize
Deserializer in little endian format. This trait can be derived if all fields of a struct implement CanonicalDeserialize and the derive feature is enabled.

CanonicalDeserializeWithFlags
Deserializer in little endian format allowing flags to be encoded.

CanonicalSerialize
Serializer in little endian format. This trait can be derived if all fields of a struct implement CanonicalSerialize and the derive feature is enabled.

CanonicalSerializeWithFlags
Serializer in little endian format allowing to encode flags.

ConstantSerializedSize
Helper trait to get serialized size for constant sized structs.

Flags

Functions

buffer_bit_byte_size

buffer_byte_size