Trait algebra_core::serialize::CanonicalSerialize

pub trait CanonicalSerialize {
    fn serialize<W: Write>(&self, writer: W) -> Result<(), SerializationError>;

    fn serialized_size(&self) -> usize;

    fn serialize_uncompressed<W: Write>(
        &self, 
        writer: W
    ) -> Result<(), SerializationError> { ... }

    fn serialize_unchecked<W: Write>(
        &self, 
        writer: W
    ) -> Result<(), SerializationError> { ... }

    fn uncompressed_size(&self) -> usize { ... }
}

Serializer in little endian format. This trait can be derived if all fields of a struct implement CanonicalSerialize and the derive feature is enabled.

Example

// The `derive` feature must be set for the derivation to work.
use algebra_core::serialize::*;

#[derive(CanonicalSerialize)]
struct TestStruct {
    a: u64,
    b: (u64, (u64, u64)),
}

If your code depends on algebra instead, the example works analogously when importing algebra::serialize::*.

Required methods

fn serialize<W: Write>(&self, writer: W) -> Result<(), SerializationError>

Serializes self into writer.

fn serialized_size(&self) -> usize

Loading content...
Provided methods

fn serialize_uncompressed<W: Write>(
&self,
writer: W
) -> Result<(), SerializationError>

Serializes self into writer without compression.

fn serialize_unchecked<W: Write>(
&self,
writer: W
) -> Result<(), SerializationError>

Serializes self into writer without compression, and without performing validity checks. Should be used only when there is no danger of adversarial manipulation of the output.

fn uncompressed_size(&self) -> usize

Loading content...